# docs/swagger.yaml
basePath: /
definitions:
...
  api.adminStoryConfigRequest:
    properties:
      levelRequired:
        type: integer
    type: object
...
  api.storiesResponse:
    properties:
      hasMore:
        description: |-
          HasMore says whether another page exists — what an infinite-scrolling
          client actually needs, so it never has to do the arithmetic.
        type: boolean
      limit:
        description: Limit is the window size actually applied; omitted when unpaginated.
        type: integer
      page:
        description: Page is 1-based; omitted when the caller asked for no page.
        type: integer
      stories:
        items:
          $ref: '#/definitions/models.Story'
        type: array
      total:
        description: Total is the row count across every page.
        type: integer
    type: object
...
  httpx.ErrorBody:
    properties:
      error:
        $ref: '#/definitions/httpx.ErrorDetail'
    type: object
  httpx.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
      request_id:
        type: string
    type: object
...
  models.Image:
    properties:
      createdAt:
        type: string
      height:
        type: integer
      id:
        type: string
      isChallenge:
        type: boolean
      publicId:
        type: string
      thumbUrl:
        description: |-
          ThumbURL delivers the same picture resized and re-encoded for a list
          row or grid tile. Clients showing a small copy should read this instead
          of URL: it is typically a twentieth of the bytes.
        type: string
      title:
        type: string
      url:
        type: string
      width:
        type: integer
    type: object
...
  models.Story:
    properties:
      author:
        type: string
      category:
        type: string
      completedPages:
        description: for the requesting user
        type: integer
      coverThumbUrl:
        description: CoverThumbURL is [CoverURL] sized for the library's cover cards.
        type: string
      coverUrl:
        type: string
      createdAt:
        type: string
      description:
        type: string
      earnedXp:
        description: xp earned from this story
        type: integer
      id:
        type: string
      levelRequired:
        description: |-
          LevelRequired is the player level needed before the story can be played.
          1 means open to everyone.
        type: integer
      title:
        type: string
      totalPages:
        type: integer
      unlocked:
        description: for the requesting user
        type: boolean
    type: object
...
  models.StoryPage:
    properties:
      completed:
        description: for the requesting user
        type: boolean
      description:
        type: string
      gridCols:
        type: integer
      gridRows:
        type: integer
      id:
        type: string
      image:
        $ref: '#/definitions/models.Image'
      index:
        type: integer
      playSeconds:
        type: integer
      storyId:
        type: string
      title:
        type: string
      unlocked:
        description: for the requesting user
        type: boolean
      xpReward:
        type: integer
    type: object
...
info:
  contact: {}
  description: |-
    Backend for the Jigsaw Puzzle app: email/password auth, image
    collection, admin daily challenges (Cloudinary) and per-day
    reputation ("Danh vọng").
  title: Jigsaw Puzzle API
  version: "1.0"
paths:
...
  /api/v1/admin/stories:
    get:
      parameters:
      - description: Page number (1-based); needs limit or defaults to 8
        in: query
        name: page
        type: integer
      - description: Rows per page; omit for the whole list
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.storiesResponse'
      security:
      - BearerAuth: []
      summary: List stories (admin)
      tags:
      - admin
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Cover image (max 15MB)
        in: formData
        name: file
        required: true
        type: file
      - description: Story title
        in: formData
        name: title
        required: true
        type: string
      - description: Category
        in: formData
        name: category
        type: string
      - description: Author
        in: formData
        name: author
        type: string
      - description: Description
        in: formData
        name: description
        type: string
      - description: Player level needed to play it (default 1)
        in: formData
        name: levelRequired
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Story'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create story (admin)
      tags:
      - admin
  /api/v1/admin/stories/{id}/config:
    put:
      consumes:
      - application/json
      parameters:
      - description: Story id (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Story configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.adminStoryConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Story'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update story config (admin)
      tags:
      - admin
  /api/v1/admin/stories/{id}/pages:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Story id (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Page image (max 15MB)
        in: formData
        name: file
        required: true
        type: file
      - description: Chapter/page title
        in: formData
        name: title
        type: string
      - description: Narrative text
        in: formData
        name: description
        type: string
      - description: Grid columns (default 4)
        in: formData
        name: gridCols
        type: integer
      - description: Grid rows (default 6)
        in: formData
        name: gridRows
        type: integer
      - description: Time limit in seconds (default 300)
        in: formData
        name: playSeconds
        type: integer
      - description: XP awarded (default 100)
        in: formData
        name: xpReward
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.StoryPage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
      security:
      - BearerAuth: []
      summary: Add story page (admin)
      tags:
      - admin
  /api/v1/admin/stories/{id}/pages/{pageId}:
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: Story id (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Page id (UUID)
        in: path
        name: pageId
        required: true
        type: string
      - description: New image (optional)
        in: formData
        name: file
        type: file
      - description: Chapter/page title
        in: formData
        name: title
        type: string
      - description: Narrative text
        in: formData
        name: description
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update story page (admin)
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the session token issued by
      /api/v1/auth/login or /register.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"